+++ Pandora/Paradigm/Primary/Object.hs
@@ -3,7 +3,7 @@
 module Pandora.Paradigm.Primary.Object (module Exports) where
 
 import Pandora.Paradigm.Primary.Object.Denumerator as Exports
-import Pandora.Paradigm.Primary.Object.Natural as Exports
+import Pandora.Paradigm.Primary.Object.Numerator as Exports
 import Pandora.Paradigm.Primary.Object.Ordering as Exports
 import Pandora.Paradigm.Primary.Object.Boolean as Exports