+++ Pandora/Paradigm/Structure/Interface/Set.hs
@@ -10,7 +10,7 @@ import Pandora.Paradigm.Primary.Functor.Maybe (Maybe (Nothing))
 import Pandora.Paradigm.Primary.Functor.Predicate (equate)
 import Pandora.Paradigm.Primary.Functor.Product (attached)
 import Pandora.Paradigm.Primary.Object.Boolean (Boolean (True, False))
-import Pandora.Paradigm.Primary.Object.Natural (Natural (Zero))
+import Pandora.Paradigm.Primary.Object.Numerator (Numerator (Zero))
 import Pandora.Paradigm.Structure.Ability.Monotonic (Monotonic (reduce), find)
 import Pandora.Paradigm.Inventory.State (State, modify)
 import Pandora.Paradigm.Controlflow.Effect (run)
@@ -21,5 +21,5 @@ member x = reduce @a @(Maybe a) (\_ _ -> True) False . find (equate x)
 subset :: (Monotonic a (t a), Traversable t, Setoid a, Setoid (t a)) => t a -> t a -> Boolean
 subset ss s = Nothing /= (ss ->> find % s . equate)
 
-cardinality :: Traversable t => t a -> Natural
-cardinality s = attached . run @(State _) % Zero $ s ->> (!) (modify @Natural (+ one))
+cardinality :: Traversable t => t a -> Numerator
+cardinality s = attached . run @(State _) % Zero $ s ->> (!) (modify @Numerator (+ one))